Implement `teams create` as plugin
